Add files via upload
FDR (False Discovery Rate)is a statistical method used in multiple hypothesis testing to correct for multiple comparisons and control the expected proportion of false positives among the rejected hypotheses. In the context of pathway analysis, FDR is used to adjust p-values of the enriched pathways to account for the multiple testing problem, as you are testing multiple pathways simultaneously.